# src/main/resources/static/api/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Law Firm API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080/"
    }
  ],
  "paths": {
    "/api/v1/clients": {
      "get": {
        "tags": [
          "clients collection"
        ],
        "summary": "List clients (paginateds)",
        "parameters": [
          {
            "name": "pageNumber",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "required": false
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "required": false
          },
          {
            "name": "searchTerm",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "required": false,
            "description": "Searches cpf or fullName"
          },
          {
            "name": "benefitType",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "required": false,
            "style": "form",
            "explode": true,
            "description": "Filter by benefit types"
          },
          {
            "name": "situation",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "required": false,
            "style": "form",
            "explode": true,
            "description": "Filter by situation"
          },
          {
            "name": "createdFrom",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "required": false,
            "description": "Start instant (inclusive) for createdAt filter. Accepts ISO-8601 date-time (e.g. 2026-02-01T00:00:00.000Z). Legacy MM-DD-YYYY is still supported for backward compatibility."
          },
          {
            "name": "createdTo",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "required": false,
            "description": "End instant (inclusive) for createdAt filter. Accepts ISO-8601 date-time (e.g. 2026-02-07T00:00:00.000Z). Legacy MM-DD-YYYY is still supported for backward compatibility."
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseClients"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "clients collection"
        ],
        "summary": "Create client",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientRequest"
              },
              "example": {
                "fullName": "Maria da Silva",
                "birthDate": "1985-06-15",
                "cpf": "123.456.789-00",
                "motherName": "Ana Silva",
                "mobilePhone": "+55 11 99999-0000",
                "inssPassword": "senhaINSS123",
                "gender": "Feminino",
                "rg": "MG-12.345.678",
                "email": "maria.silva@example.com",
                "referencePhone": "+55 11 98888-0000",
                "referenceResponsible": "João Silva",
                "maritalStatus": "Casado(a)",
                "benefit": "Aposentadoria por idade",
                "situation": "formulário preenchido",
                "beneficiaryNumber": "BN-2026-0001",
                "nitPis": "NIT12345678",
                "profession": "Professor",
                "ctps": "CTPS-12345",
                "ctpsSeries": "S1",
                "contributionTime": 35,
                "nonBillable": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseClient"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - validation or business rule",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseClient"
                },
                "examples": {
                  "errorEnvelope": {
                    "value": {
                      "success": false,
                      "data": [],
                      "errors": [
                        {
                          "field": "string",
                          "message": "string",
                          "code": "VALIDATION_ERROR"
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Conflict - data integrity or duplicate",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseClient"
                },
                "examples": {
                  "errorEnvelope": {
                    "value": {
                      "success": false,
                      "data": [],
                      "errors": [
                        {
                          "field": "string",
                          "message": "string",
                          "code": "DUPLICATE_RECORD"
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseClient"
                },
                "examples": {
                  "errorEnvelope": {
                    "value": {
                      "success": false,
                      "data": [],
                      "errors": [
                        {
                          "field": null,
                          "message": "Erro interno do sistema",
                          "code": "SYSTEM_ERROR"
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/clients/{id}": {
      "get": {
        "tags": [
          "clients collection"
        ],
        "summary": "Get client by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseClient"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseClient"
                },
                "examples": {
                  "validationError": {
                    "value": {
                      "success": false,
                      "data": [],
                      "errors": [
                        {
                          "field": "id",
                          "message": "ID inválido",
                          "code": "VALIDATION_ERROR"
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseClient"
                },
                "examples": {
                  "notFound": {
                    "value": {
                      "success": false,
                      "data": [],
                      "errors": [
                        {
                          "field": null,
                          "message": "Recurso não encontrado",
                          "code": "NOT_FOUND"
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseClient"
                },
                "examples": {
                  "systemError": {
                    "value": {
                      "success": false,
                      "data": [],
                      "errors": [
                        {
                          "field": null,
                          "message": "Erro interno do sistema",
                          "code": "SYSTEM_ERROR"
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "clients collection"
        ],
        "summary": "Update client by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseClientPatch"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - validation or business rule",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseClient"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseClient"
                }
              }
            }
          },
          "409": {
            "description": "Conflict - data integrity or duplicate",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseClient"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseClient"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "clients collection"
        ],
        "summary": "Patch client (partial update of situation and/or nonBillable)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientPatchRequest"
              },
              "example": {
                "situation": "planejamento concluído",
                "nonBillable": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseClient"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - validation or business rule",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseClient"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseClient"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseClient"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "clients collection"
        ],
        "summary": "Delete client by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseClient"
                }
              }
            }
          },
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseClient"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseClient"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseClient"
                }
              }
            }
          }
        }
      }
    }
...
    "/api/v1/clients/{id}/history": {
      "get": {
        "tags": [
          "clients collection"
        ],
        "summary": "Get situation history for a client (paginated, newest first)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "pageNumber",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 1 },
            "description": "1-based page number (default: 1)"
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 10 },
            "description": "Number of items per page (default: 10)"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiResponseClientHistory" }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiResponseClientHistory" }
              }
            }
          }
        }
      }
    }
  },
...
  "components": {
    "schemas": {
...
      "ClientListItem": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "fullName": { "type": "string" },
          "cpf": { "type": "string" },
          "mobilePhone": { "type": "string" },
          "benefit": { "type": "string", "description": "Label from BenefitType enum" },
          "situation": { "type": "string", "description": "Label from Situation enum" },
          "beneficiaryNumber": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time", "readOnly": true }
        }
      },
      "ClientRequest": {
        "type": "object",
        "required": [
          "fullName",
          "birthDate",
          "cpf",
          "motherName",
          "mobilePhone",
          "inssPassword",
          "gender",
          "benefit",
          "situation"
        ],
        "properties": {
          "fullName": {
            "type": "string"
          },
          "birthDate": {
            "type": "string",
            "format": "date"
          },
          "cpf": {
            "type": "string"
          },
          "motherName": {
            "type": "string"
          },
          "mobilePhone": {
            "type": "string"
          },
          "inssPassword": {
            "type": "string"
          },
          "gender": {
            "type": "string"
          },
          "rg": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "referencePhone": {
            "type": "string"
          },
          "referenceResponsible": {
            "type": "string"
          },
          "maritalStatus": {
            "type": "string"
          },
          "benefit": {
            "type": "string"
          },
          "situation": {
            "type": "string"
          },
          "beneficiaryNumber": {
            "type": "string"
          },
          "nitPis": {
            "type": "string"
          },
          "profession": {
            "type": "string"
          },
          "ctps": {
            "type": "string"
          },
          "ctpsSeries": {
            "type": "string"
          },
          "contributionTime": {
            "type": "integer"
          },
          "nonBillable": {
            "type": "boolean",
            "default": false
          }
        }
      },
      "ClientUpdateRequest": {
        "type": "object",
        "description": "Fields that can be updated via PUT (exclude read-only/generated fields: id, firstName, lastName, createdAt, updatedAt, age)",
        "required": [
          "fullName",
          "birthDate",
          "cpf",
          "motherName",
          "mobilePhone",
          "inssPassword",
          "gender"
        ],
        "properties": {
          "fullName": {
            "type": "string"
          },
          "birthDate": {
            "type": "string",
            "format": "date"
          },
          "cpf": {
            "type": "string"
          },
          "motherName": {
            "type": "string"
          },
          "mobilePhone": {
            "type": "string"
          },
          "inssPassword": {
            "type": "string"
          },
          "gender": {
            "type": "string",
            "description": "Allowed values are the labels from the Gender enum (Masculino, Feminino, Não-binário, Outro)"
          },
          "rg": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "referencePhone": {
            "type": "string"
          },
          "referenceResponsible": {
            "type": "string"
          },
          "maritalStatus": {
            "type": "string",
            "description": "Labels from MaritalStatus enum (Solteiro(a), Casado(a), ... )"
          },
          "benefit": {
            "type": "string",
            "description": "Labels from BenefitType enum"
          },
          "situation": {
            "type": "string",
            "description": "Labels from Situation enum"
          },
          "beneficiaryNumber": {
            "type": "string"
          },
          "nitPis": {
            "type": "string"
          },
          "profession": {
            "type": "string"
          },
          "ctps": {
            "type": "string"
          },
          "ctpsSeries": {
            "type": "string"
          },
          "contributionTime": {
            "type": "integer"
          },
          "nonBillable": {
            "type": "boolean",
            "default": false
          }
        }
      },
      "ClientPatchRequest": {
        "type": "object",
        "properties": {
          "situation": {
            "type": "string",
            "description": "Label from Situation enum"
          },
          "nonBillable": {
            "type": "boolean"
          }
        }
      },
      "ApiError": {
        "type": "object",
        "properties": {
          "field": {
            "type": "string",
            "nullable": true
          },
          "message": {
            "type": "string"
          },
          "code": {
            "type": "string"
          }
        },
        "required": [
          "message",
          "code"
        ]
      },
      "ApiResponseClients": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ClientListItem"
            }
          },
          "pagination": {
            "type": "object",
            "properties": {
              "pageNumber": {
                "type": "integer"
              },
              "pageSize": {
                "type": "integer"
              },
              "totalRecords": {
                "type": "integer"
              },
              "totalPages": {
                "type": "integer"
              },
              "hasNextPage": {
                "type": "boolean"
              },
              "hasPreviousPage": {
                "type": "boolean"
              }
            }
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApiError"
            },
            "nullable": true
          }
        }
      },
      "ApiResponseClient": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/Client"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApiError"
            },
            "nullable": true
          }
        }
      }
...
      "Pagination": {
        "type": "object",
        "properties": {
          "pageNumber": { "type": "integer" },
          "pageSize": { "type": "integer" },
          "totalRecords": { "type": "integer" },
          "totalPages": { "type": "integer" },
          "hasNextPage": { "type": "boolean" },
          "hasPreviousPage": { "type": "boolean" }
        }
      },
      "ClientPatchResponse": {
        "type": "object",
        "properties": {
          "message": { "type": "string" }
        }
      },
      "ApiResponseClientPatch": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "data": { "$ref": "#/components/schemas/ClientPatchResponse" },
          "errors": { "type": "array", "items": { "$ref": "#/components/schemas/ApiError" } }
        }
      },
      "ClientSituationHistory": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "currentSituation": { "type": "string", "description": "Label from Situation enum" },
          "changedAt": { "type": "string", "format": "date-time" },
          "changedBy": { "type": "integer", "nullable": true }
        }
      },
...
      "ApiResponseClientHistory": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "data": { "type": "array", "items": { "$ref": "#/components/schemas/ClientSituationHistory" } },
          "pagination": { "$ref": "#/components/schemas/Pagination" },
          "errors": { "type": "array", "items": { "$ref": "#/components/schemas/ApiError" } }
        }
      }
    }
  }
}